# application/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Programmer Notebook
  version: 1.0.0
servers:
  - url: https://0.0.0.0:8080/api/v2
...
paths:
...
  /teams/{id}:
    get:
      tags:
        - teams
      summary: Get team
      description: Получение одной команды
      operationId: getTeam

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID команды
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: Success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/TeamFullInfo'
        "401":
          description: Authorization error
        "403":
          description: Недостаточно прав
        "404":
          description: No such user
        "500":
          description: Internal error

    delete:
      tags:
        - teams
      summary: Delete team
      description: Удаление команды
      operationId: deleteTeam

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID команды
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: Success
        "401":
          description: Authorization error
        "403":
          description: Недостаточно прав
        "404":
          description: No such command
        "500":
          description: Internal error

    patch:
      tags:
        - teams
      summary: Update team info
      description: Изменение данных команды
      operationId: updateTeam

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID команды
          required: true
          schema:
            type: integer

      requestBody:
        description: Новые данные команды
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/TeamInfo'
        required: true

      responses:
        "200":
          description: Success
        "401":
          description: Authorization error
        "403":
          description: Недостаточно прав
        "404":
          description: No such user
        "500":
          description: Internal error
...
  /notes/{id}:
    get:
      tags:
        - notes
      summary: Get a note
      description: Получить записку по ID или названию.
      operationId: getNote

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID записки.
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: Успешный ответ с данными записки.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/NoteFullData'
        "404":
          description: Записка не найдена.
        "500":
          description: Внутренняя ошибка сервера.

    delete:
      tags:
        - notes
      summary: Delete a note
      description: Удалить записку по ID или названию.
      operationId: deleteNote

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID записки
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: Записка успешно удалена.
        "404":
          description: Записка не найдена.
        "500":
          description: Внутренняя ошибка сервера.

    patch:
      tags:
        - notes
      summary: Update a note
      description: Обновить информацию о записке.
      operationId: updateNote

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID записки
          required: true
          schema:
            type: integer

      requestBody:
        description: Обновленные данные записки.
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/NoteFullData'

      responses:
        "200":
          description: Записка успешно обновлена.
        "404":
          description: Записка не найдена.
        "500":
          description: Внутренняя ошибка сервера.

  #####################################################################
  # Collection API
  #####################################################################
...
  /sections:
    get:
      tags:
        - sections
      summary: Get all sections
      operationId: getAllSections

      security:
        - JWTAuth: []

      responses:
        "200":
          description: Успешное получение всех разделов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SectionInfo'
        "403":
          description: Недостаточно прав
        "500":
          description: Внутренняя ошибка сервера

    post:
      tags:
        - sections
      summary: Add new section
      operationId: addSection

      security:
        - JWTAuth: []

      requestBody:
        description: Данные для добавления раздела.
        required: true
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/SectionInfo'

      responses:
        "200":
          description: Успешное добавление раздела
          content:
            'application/json':
              schema:
                type: object
                properties:
                  secID:
                    type: integer
        "403":
          description: Недостаточно прав
        "500":
          description: Внутренняя ошибка сервера
  /sections/{id}:
    get:
      tags:
        - sections
      summary: Get section by ID
      description: Получить информацию о разделе по его ID.
      operationId: getSection

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID раздела.
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: Успешное получение раздела
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionFullInfo'
        "404":
          description: Раздел не найден
        "400":
          description: Неверный запрос (например, некорректные параметры)
        "500":
          description: Внутренняя ошибка сервера

    delete:
      tags:
        - sections
      summary: Delete a section
      description: Удалить раздел по ID.
      operationId: deleteSection

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID раздела.
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: Успешное удаление раздела
        "403":
          description: Недостаточно прав
        "404":
          description: Раздел не найден
        "500":
          description: Внутренняя ошибка сервера
  /sections/{id}/notes:
    get:
      tags:
        - sections
      summary: Get all notes in a section
      description: Получить все записки, относящиеся к указанному разделу.
      operationId: getAllNotesInSection

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID раздела.
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: Успешное получение записок из раздела
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NoteInfo'
        "403":
          description: Недостаточно прав
        "404":
          description: Раздел не найден
        "500":
          description: Внутренняя ошибка сервера

    post:
      tags:
        - sections
      summary: Add a note to a section
      description: Добавление записки в раздел.
      operationId: addNoteToSec

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID коллекции
          required: true
          schema:
            type: integer
            example: 10

      requestBody:
        description: ID записки
        content:
          'application/json':
            schema:
              type: object
              properties:
                noteID:
                  type: integer
                  description: ID записки
        required: true

      responses:
        "200":
          description: Успешное добавление записки в раздел
        "403":
          description: Недостаточно прав
        "404":
          description: Раздел или записка не найдены
        "500":
          description: Внутренняя ошибка сервера
...
components:
  schemas:
...
    TeamFullInfo:
      type: object
      required:
        - id
        - name
        - registrationDate
      properties:
        id:
          type: integer
          format: int
          example: 10
        name:
          type: string
          example: Фронтенд-команда
        registrationDate:
          type: string
          example: 01.01.2001
    TeamInfo:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int
          example: 10
        name:
          type: string
          example: Фронтенд-команда

    #####################################################################
    # Note schemas
    #####################################################################
    NoteInfo:
      type: object
      required:
        - id
        - name
        - likes
        - dislikes
      properties:
        id:
          type: integer
          format: int
          example: 10
        name:
          type: string
          description: Название Записки
          example: Распраллеливание в Python
        likes:
          type: integer
          format: int
          example: 10
        dislikes:
          type: integer
          format: int
          example: 10
    NoteFullInfo:
      type: object
      required:
        - id
        - name
        - access
        - likes
        - dislikes
        - registrationDate
        - ownerID
        - sectionID
      properties:
        id:
          type: integer
          format: int
          example: 10
        access:
          type: integer
          description: Закрытая или общедоступная Записка
          format: int
          example: 1
        name:
          type: string
          description: Название Записки
          example: Распраллеливание в Python
        likes:
          type: integer
          format: int
          example: 10
        dislikes:
          type: integer
          format: int
          example: 10
        registrationDate:
          type: string
          example: 01.01.2001
        ownerID:
          type: integer
          description: ID пользователя-владельца
          format: int
          example: 10
        sectionID:
          type: integer
          description: "Если Записка находится в разделе, то здесь ID раздела, иначе 0"
          format: int
          example: 10
    NoteFullData:
      type: object
      required:
        - info
        - textFile
        - imgFile
        - otherFile
        - txtFileType
        - imgFileType
        - otherFileType
      properties:
        info:
          $ref: '#/components/schemas/NoteFullInfo'
        textFile:
          type: string
          format: binary
          description: Текстовое содержимое записки
        image:
          type: string
          format: binary
          description: Изображение (jpg или png)
        otherFile:
          type: string
          format: binary
          description: Файл неизвестного формата
        txtFileType:
          type: string
          example: "txt"
        imgFileType:
          type: string
          example: "txt"
        otherFileType:
          type: string
          example: "txt"

    #####################################################################
    # Collection schemas
    #####################################################################
...
    SectionInfo:
      type: object
      required:
        - id
        - teamID
      properties:
        id:
          type: integer
          format: int
          example: 10
        teamID:
          type: integer
    SectionFullInfo:
      type: object
      required:
        - id
        - registrationDate
        - teamID
      properties:
        id:
          type: integer
          format: int
          example: 10
        registrationDate:
          type: string
          example: 01.01.2001
        teamID:
          type: integer
...
  securitySchemes:
    JWTAuth:
      type: apiKey
      name: token
      in: header